Rename reviewService.js
